Implement Initial Data Service
This commit includes the implementation of a Data Service application using Flask. This Data Service allows the fetching, streaming, and listing of available S3 bucket data. In addition, AWS credentials are stored in a .env file and used to interact with the AWS S3 bucket. Docker configurations for running the application are also included.